move UmbPropertyEditorElement
