+++ libs/extensions-registry/index.ts
@@ -1,2 +1,3 @@
+export * from './interfaces';
 export * from './models';
 export * from './entry-point-extension-initializer';
