+++ libs/extensions-registry/interfaces/index.ts
