+++ src/backoffice/shared/property-editors/uis/document-picker/property-editor-ui-document-picker.element.ts
@@ -1,7 +1,7 @@
 import { html } from 'lit';
 import { customElement, property, state } from 'lit/decorators.js';
 import type { UmbInputDocumentPickerElement } from '../../../components/input-document-picker/input-document-picker.element';
-import { UmbPropertyEditorElement } from '@umbraco-cms/backoffice/property-editor';
+import { UmbPropertyEditorElement } from '@umbraco-cms/backoffice/extensions-registry';
 import { UmbLitElement } from '@umbraco-cms/internal/lit-element';
 import '../../../components/input-document-picker/input-document-picker.element';
 import type { DataTypePropertyPresentationModel } from '@umbraco-cms/backoffice/backend-api';
