+++ src/backoffice/shared/property-editors/uis/number-range/property-editor-ui-number-range.element.ts
@@ -2,7 +2,7 @@ import { html } from 'lit';
 import { UUITextStyles } from '@umbraco-ui/uui-css/lib';
 import { customElement, property, state } from 'lit/decorators.js';
 import type { UmbInputNumberRangeElement } from '../../../../shared/components/input-number-range/input-number-range.element';
-import { UmbPropertyEditorElement } from '@umbraco-cms/backoffice/property-editor';
+import { UmbPropertyEditorElement } from '@umbraco-cms/backoffice/extensions-registry';
 import { UmbLitElement } from '@umbraco-cms/internal/lit-element';
 import '../../../../shared/components/input-number-range/input-number-range.element';
