+++ src/backoffice/shared/property-editors/uis/upload-field/property-editor-ui-upload-field.element.ts
@@ -3,7 +3,7 @@ import { UUITextStyles } from '@umbraco-ui/uui-css/lib';
 import { customElement, property, state } from 'lit/decorators.js';
 import { UmbInputUploadFieldElement } from '../../../../shared/components/input-upload-field/input-upload-field.element';
 import type { DataTypePropertyPresentationModel } from '@umbraco-cms/backoffice/backend-api';
-import type { UmbPropertyEditorElement } from '@umbraco-cms/backoffice/property-editor';
+import type { UmbPropertyEditorElement } from '@umbraco-cms/backoffice/extensions-registry';
 import { UmbLitElement } from '@umbraco-cms/internal/lit-element';
 
 /**
